src/ai-service/requirements.txt
fastapi
uvicorn
opentelemetry-distro
opentelemetry-exporter-otlp
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-logging
psycopg2-binary
pika
textblob